refactor(cli): isolate egress engine contract
